Initial pass at job state.
This uses the reactor context state to hold the
mapping of job ids to messages. It appears to be
the most publicly exposed method of storing state
since writing directly to files involves elevated
permissions. More notes on that below...

Now relies on the posix path to a given project to
handle input files. It is plausible that this could
shift to using the agavepy API to download a file
from a given agave URI. This is specifically for
input of data though since the download API has a
synchronous call.

Writing to disk is a bit more involved without having
special permissions. It appears that agavepy allows a
file handle to be uploaded to a target agave URI but
it only does so in an async fashion where the callback
is a provided URL. I do not know what the implications
are for async writes and reactor execution life time.
e.g. if an execution is in the middle of a file upload
but the main thread of the execution finishes then will
the execution block until the upload completes? If it
does not block then we risk losing output data that
has taken longer than the execution to write.

Maybe there is a synchronous method somewhere in the
API.

Metadata is also a potential option. It seems to have
a synchronous API that can associate arbitrary JSON
with a specific TACC UUID.