enable the handling of the job completion messages
The reactor will now record any launched job and
associate the job id with the original message
dictionary. The job is also constructed such that
it sends a FINISHED notification back to the
reactor. The reactor will see this message and
pop the message dictionary associated with the
job id from the state and with run the finalize
pass on the message.

TODO: the reactor should also handle events like
FAILED. Its a pretty simple addition but not yet
added.