Setup file uploads from the coordinate reactor
Note:
It has worked for small files. I do not know if it will encounter
any errors if used with large files. It depends on if there are
any controls in place for how long reactors can live and if the
call is async on the local end. We don't actually listen for an
upload completed callback. We just fire and forget (and exit the
reactor).