update make and docker files to pull in dependencies
Its possible I pulled in more than necessary here. I need to
go through and double check. This is what was needed to make
it work within docker for me.

I also shifted the output mount to point to to a bind mount
instead of a volume (using the newer mount commands). It was
mounting as a volume before which is fine once I finish my
testing. But for now output is written to the bind mount.