add xplan design to the coordinator
The coordinator can now call the xplan design app. This starts a
job that upon completion calls back to the coordinator and calls
the finalize_message function in xplandesignmessage.py. Here it
just prints out the archive path where the job dumped the results
of the app. It could either download from there or the app could
be adjusted to upload them to a path other than the archive path.