Connect xplan design message to new file functions
This will download_dir the challenge_problem dir from the archived
output of the xplan design app's job into the coordinate reactor's
scratch space. From here the reactor can operate on the results of
the download on the local system. Upon completion the reactor then
runs upload_dir to the challenge_problem dir in the out_dir found
in the original message dictionary.

Also updates the upload_dir function to ensure the directory for
the challenge_problem exists in the out_dir agave path.