extra handling of job response messages.
This should help with debugging if a job sends a message with an
unexpected status (like BLOCKED). We are not actually registering
for those messages but I have seen them arrive after a job ended
up BLOCKED. So it is possible that the webhook handlers need some
extra code to correct the execution if a non-FINISHED status msg
arrives.