Apply state diff as patch in finalize step of reactor
This _might_ be enough to ensure the state.json does not
get corrupted. Though it is possible we will need to seed
the initial state with the empty properties to ensure the
state diff picked up the expected changes.

Alternatively we will need to go with a deep merge where
we write a merge step for each of the properties. It is
less general but also more likely to work as intended.